+++ src/entities/ProposalSurveyTopics/types.ts
